+++ apps/api/package.json
@@ -18,7 +18,6 @@
     "@sentry/tracing": "^7.43.0",
     "cors": "^2.8.5",
     "dedent": "^0.7.0",
-    "dotenv": "^16.3.1",
     "express": "^4.18.3",
     "helmet": "^6.0.1",
     "ts-pattern": "4.3.0",
