+++ packages/core/src/modules/employment/use-cases/save-company-if-necessary.test.ts
@@ -1,3 +1,5 @@
+import { describe, test, expect, spyOn } from 'bun:test';
+
 import { company1, company4 } from '@oyster/db/test/constants';
 
 import { db } from '@/infrastructure/database';
@@ -22,16 +24,14 @@ describe(saveCompanyIfNecessary.name, () => {
   });
 
   test('If the company does not exist in our database, should call the Crunchbase API.', async () => {
-    const mock = vi
-      .spyOn(module, 'getCrunchbaseOrganization')
-      .mockResolvedValue({
-        crunchbaseId: company4.crunchbaseId,
-        description: company4.description!,
-        domain: company4.domain!,
-        imageUrl: company4.imageUrl!,
-        name: company4.name,
-        stockSymbol: company4.stockSymbol!,
-      });
+    const mock = spyOn(module, 'getCrunchbaseOrganization').mockResolvedValue({
+      crunchbaseId: company4.crunchbaseId,
+      description: company4.description!,
+      domain: company4.domain!,
+      imageUrl: company4.imageUrl!,
+      name: company4.name,
+      stockSymbol: company4.stockSymbol!,
+    });
 
     await db.transaction().execute(async (trx) => {
       return saveCompanyIfNecessary(trx, company4.crunchbaseId);
@@ -41,7 +41,7 @@ describe(saveCompanyIfNecessary.name, () => {
   });
 
   test('If the company does not exist in our database, it should create it from the Crunchbase API result.', async () => {
-    vi.spyOn(module, 'getCrunchbaseOrganization').mockResolvedValue({
+    spyOn(module, 'getCrunchbaseOrganization').mockResolvedValue({
       crunchbaseId: company4.crunchbaseId,
       description: company4.description!,
       domain: company4.domain!,
