+++ packages/core/tsconfig.json
@@ -8,7 +8,6 @@
     "moduleResolution": "Bundler",
     "paths": {
       "@/*": ["./src/*"]
-    },
-    "types": ["vitest/globals"]
+    }
   }
 }
