+++ packages/utils/package.json
@@ -7,12 +7,11 @@
   "scripts": {
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
-    "test": "vitest run",
+    "test": "bun test",
     "type-check": "tsc --noEmit"
   },
   "devDependencies": {
     "@oyster/eslint-config": "workspace:*",
-    "nanoid": "^3.0.0",
-    "vitest": "^1.3.1"
+    "nanoid": "^3.0.0"
   }
 }
