+++ packages/utils/src/get-cookie.test.ts
@@ -1,3 +1,5 @@
+import { describe, expect, test } from 'bun:test';
+
 import { getCookie } from './get-cookie';
 
 describe(getCookie.name, () => {
