+++ packages/utils/src/sleep.test.ts
@@ -1,3 +1,5 @@
+import { describe, expect, test } from 'bun:test';
+
 import { sleep } from './sleep';
 
 describe(sleep.name, () => {
